implemented frozenset conversion
