+++ woltka/tests/test_classify.py
@@ -14,8 +14,7 @@ from shutil import rmtree
 from tempfile import mkdtemp
 
 from woltka.classify import (
-    assign_none, assign_free, assign_rank, count, count_strata, majority,
-    strip_index, demultiplex)
+    assign_none, assign_free, assign_rank, count, count_strata, majority)
 
 
 class ClassifyTests(TestCase):
@@ -164,62 +163,6 @@ class ClassifyTests(TestCase):
         self.assertIsNone(obs)
         obs = majority([])
         self.assertIsNone(obs)
-
-    def test_strip_index(self):
-        subs = [{'G1_1', 'G1_2', 'G2_3', 'G3'},
-                {'G1_1', 'G1.3', 'G4_5', 'G4_x'}]
-        obs = strip_index(subs)
-        exp = [{'G1', 'G2', 'G3'},
-               {'G1', 'G1.3', 'G4', 'G4'}]
-        self.assertListEqual(list(obs), exp)
-
-        subs = [{'NC_123456.1_300', 'ABCD000001.20_101'}]
-        obs = strip_index(subs)
-        exp = [{'NC_123456.1', 'ABCD000001.20'}]
-        self.assertListEqual(list(obs), exp)
-
-        subs = [{'G1.1', 'G1.2', 'G2'},
-                {'G1.1', 'G1.3', 'G3_x'}]
-        obs = strip_index(subs, sep='.')
-        exp = [{'G1', 'G2'},
-               {'G1', 'G3_x'}]
-        self.assertListEqual(list(obs), exp)
-
-    def test_demultiplex(self):
-        # simple case
-        rmap = [('S1_R1', 5),
-                ('S1_R2', 12),
-                ('S1_R3', 3),
-                ('S2_R1', 10),
-                ('S2_R2', 8),
-                ('S2_R4', 7),
-                ('S3_R2', 15),
-                ('S3_R3', 1),
-                ('S3_R4', 5)]
-        obs = demultiplex(*zip(*rmap))
-        exp = {'S1': [('R1',  5),
-                      ('R2', 12),
-                      ('R3',  3)],
-               'S2': [('R1', 10),
-                      ('R2',  8),
-                      ('R4',  7)],
-               'S3': [('R2', 15),
-                      ('R3',  1),
-                      ('R4',  5)]}
-        self.assertEqual(obs.keys(), exp.keys())
-        for s in obs:
-            self.assertListEqual(list(map(tuple, obs[s])), list(zip(*exp[s])))
-
-        # change separator, no result
-        obs = demultiplex(*zip(*rmap), sep='.')
-        self.assertEqual(obs.keys(), {''})
-        self.assertListEqual(list(map(tuple, obs[''])), list(zip(*rmap)))
-
-        # enforce sample Ids
-        obs = demultiplex(*zip(*rmap), samples=['S1', 'S2', 'SX'])
-        self.assertEqual(obs.keys(), {'S1', 'S2'})
-        for s in ('S1', 'S2'):
-            self.assertListEqual(list(map(tuple, obs[s])), list(zip(*exp[s])))
 
 
 if __name__ == '__main__':
